+++ packages/wotex-lab/.formatter.exs
@@ -1,4 +1,4 @@
 [
-  inputs: ["{mix,.formatter,.check,.credo,.doctor}.exs", "{lib,test}/**/*.{ex,exs}", "bin/*.exs"],
+  inputs: ["{mix,.formatter,.check,.credo,.doctor}.exs", "{lib,test}/**/*.{ex,exs}", "bin/**/*.exs"],
   line_length: 100
 ]